refactor(codehealth): oauth, audit, desirepath, root and small packages to green
